+++ amazonorders/__init__.py
@@ -1,3 +1,3 @@
 __copyright__ = "Copyright (c) 2024 Alex Laird"
 __license__ = "MIT"
-__version__ = "2.0.0"
+__version__ = "3.0.0"